requirements.txt
# Ad Retrieval System - Core Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# ML & Embeddings
sentence-transformers[onnx]>=2.3.0
transformers>=4.37.0
torch>=2.1.0
keybert>=0.9.0

# Vector Search
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# Data Processing
numpy>=1.26.0
pandas>=2.1.0

# HTTP Client (for benchmarking)
httpx>=0.26.0
aiohttp>=3.9.0

# Image Processing
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
orjson>=3.9.0
